Split demo app to main, app and home_page
